Move AnimationTab and subordinate components into own folder.
Attempting to follow styleguide directive here:

https://github.com/airbnb/javascript/tree/master/react#naming

However, our build system won't currently let us require a directory and implicitly import index.jsx from that directory, so we have to explicitly import AnimationTab/index.jsx from GameLabView.jsx.
