+++ apps/src/gamelab/AnimationTab/AnimationSequenceListItem.jsx
@@ -1,8 +1,8 @@
 /** A single list item representing an animation sequence. */
 'use strict';
 
-var _ = require('../lodash');
-var color = require('../color');
+var _ = require('../../lodash');
+var color = require('../../color');
 var connect = require('react-redux').connect;
 var ListItemButtons = require('./ListItemButtons.jsx');
 var ListItemThumbnail = require('./ListItemThumbnail.jsx');
